Remove git command to get current branch name
Instead of making a separate syscall to get the current branch name it's
now inferring it from the formatting applied in the branch list. This
assumes that all git environments use * to indicate which branch is the
current one, if that assumption proves to be false the syscall may need
to be re-implemented.