only add the current branch once
Moving the code to add branch names into an else block so that the
current branch isn't added twice. Hacking an extra space into the
current branch name to enforce consistent vertical spacing. This is a
temporary hack and will be fixed later, most likely by stripping all
formatting from the branch names as the data comes in.